all basic cms functionality present; still need to add "enabled" toggles for all doc types / collections

+++ express-routes/controllers/ApplicationSpecific/customModuleHandlers.js
@@ -2,32 +2,32 @@ import { Schema, Types } from 'mongoose'
 import CustomModule from '../../models/ApplicationSpecific/CustomModule'
 
 const createCustomModule = (req, res) => {
-  // if (!req.session.appuserauth) res.send({ error: 'not authorized' })
-  // else {
+  if (!req.session.appuserauth) res.send({ error: 'not authorized' })
+  else {
     // console.log(req.body)
-    if (req.body.image_url) {
-      CustomModule({ image_url: req.body.image_url })
+    if (req.body.name && req.body.image_url) {
+      CustomModule({ name: req.body.name, image_url: req.body.image_url, creator_id: Types.ObjectId(req.session.appuserid) || '' })
       .save()
       .then(() => res.send({ success: 'custom module saved' }))
       .catch(err => res.send({ error: 'db err occurred' }))
     } else {
       res.send({ error: 'incorrect params provided' })
     }
-  // }
+  }
 }
 
 const viewCustomModules = (req, res) => {
-  // if (!req.session.appuserauth) res.send({ error: 'not authorized' })
-  // else {
+  if (!req.session.appuserauth) res.send({ error: 'not authorized' })
+  else {
     CustomModule.find()
       .then(data => res.send({ success: data }))
       .catch(err => res.send({ error: 'error occurred getting tour module data' }))
-  // }
+  }
 }
 
 const deleteCustomModule = (req, res) => {
-  // if (!req.session.appuserauth) res.send({ error: 'not authorized' })
-  // else {
+  if (!req.session.appuserauth) res.send({ error: 'not authorized' })
+  else {
     if (req.params.id) {
       let id = req.params.id
       CustomModule.deleteOne({ _id: id })
@@ -36,12 +36,12 @@ const deleteCustomModule = (req, res) => {
     } else {
       res.send({ error: 'no object id provided' })
     }
-  // }
+  }
 }
 
 const getOneCustomModule = (req, res) => {
-  // if (!req.session.appuserauth) res.send({ error: 'not authorized' })
-  // else {
+  if (!req.session.appuserauth) res.send({ error: 'not authorized' })
+  else {
     if (req.params.id) {
       const { id } = req.params
       CustomModule.findOne({ _id: id })
@@ -50,19 +50,20 @@ const getOneCustomModule = (req, res) => {
     } else {
       res.send({ error: 'no object id provided' })
     }
-  // }
+  }
 }
 
 const updateCustomModule = (req, res) => {
-  // if (!req.session.appuserauth) res.send({ error: 'not authorized' })
-  // else {
-    if (req.params.id) {
+  if (!req.session.appuserauth) res.send({ error: 'not authorized' })
+  else {
+    if (req.params.id && req.body.name && req.body.image_url) {
       const { id } = req.params
       console.log(id)
       CustomModule.findOne({ _id: id })
         .then(doc => {
           if (doc) {
-            doc.image_url = req.body.image_url
+            doc.name = req.body.name
+            if (req.body.image_url != '') doc.image_url = req.body.image_url
             doc.save()
               .then(doc => res.send({ success: `updated ${doc._id} successfully` }))
               .catch(err => res.send({ error: `error occurred updating ${doc._id}` }))
@@ -70,9 +71,9 @@ const updateCustomModule = (req, res) => {
         })
         .catch(err => res.send({ error: 'error occurred finding custom module doc' }))
     } else {
-      res.send({ error: 'no id param sent' })
+      res.send({ error: 'incorrect params' })
     }
-  // }
+  }
 }
 
 export {
